requirements.txt
openai
speechrecognition
python-dotenv
requests
pydub
pyaudio
pyttsx3
requests
keyboard
datetime
